backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.2.1
environs==14.1.0
fastapi==0.115.6
greenlet==3.1.1
loguru==0.7.3
passlib==1.7.4
pydantic==2.10.5
pydantic-settings==2.7.1
PyJWT==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.37
uvicorn==0.34.0
psycopg2-binary==2.9.11
opencv-python-headless==4.10.0.84
numpy==1.26.4
email-validator==2.3.0
Pillow==10.4.0
mediapipe==0.10.10
scipy==1.11.4
python-dotenv==1.0.1
torch==2.7.0+cpu